Some long-overdue reorg of the files
