+++ src/EnergyPlus/AirflowNetwork/CMakeLists.txt
@@ -6,9 +6,9 @@ if(${CMAKE_MAJOR_VERSION}.${CMAKE_MINOR_VERSION} VERSION_GREATER "3.0")
   cmake_policy(SET CMP0054 NEW) # CMake 3.1 added this policy
 endif()
 
-set(HDRS include/AirflowNetwork/Solver.hpp include/AirflowNetwork/Elements.hpp include/AirflowNetwork/Properties.hpp)
+set(HDRS include/AirflowNetwork/Types.hpp include/AirflowNetwork/Solver.hpp include/AirflowNetwork/Elements.hpp include/AirflowNetwork/Linkages.hpp include/AirflowNetwork/Properties.hpp include/AirflowNetwork/Nodes.hpp)
 
-set(SRCS src/Solver.cpp src/Elements.cpp src/Properties.cpp)
+set(SRCS src/Solver.cpp src/Elements.cpp src/Linkages.cpp src/Properties.cpp src/Nodes.cpp)
 
 # Create a static library
 # this will be linked statically to create the DLL and also the unit tests
